autenticacao funcionando

+++ gerente/conta.php
@@ -1,4 +1,7 @@
 <!DOCTYPE html>
+<?php
+    session_start();
+?>
 <html lang="pt-BR">
 <head>
     <meta charset="UTF-8">
@@ -19,12 +22,15 @@
                 $stmt->execute();
                 $gerentes = $stmt->fetchAll();
 
+                
                 echo "<tr><th>Usuário</th><th>Nome</th><th>Data de Nascimento</th><th>Sexo</th><th>E-mail</th><th>Editar</th><th>Excluir</th></tr>";
 
                 foreach($gerentes as $gerente){
-                    $editar = "<a href=cadastro.php?acao=editar&id=".$gerente["id"].">Alt</a>";
-                    $excluir = "<a href='acao.php?acao=excluir&id=".$gerente["id"]."'>Exc</a>";
-                    echo "<tr><td>".$gerente["usuario"]."</td>"."<td>".$gerente["nome"]."</td>"."<td>".$gerente["dataNasc"]."</td>"."<td>".$gerente["sexo"]."</td>"."<td>".$gerente["email"]."</td><td>$editar</td>"."<td>$excluir</td>"."</tr>";
+                    if($_SESSION['idGerente'] === $gerente['id']){
+                        $editar = "<a href=cadastro.php?acao=editar&id=".$gerente["id"].">Alt</a>";
+                        $excluir = "<a href='acao.php?acao=excluir&id=".$gerente["id"]."'>Exc</a>";
+                        echo "<tr><td>".$gerente["usuario"]."</td>"."<td>".$gerente["nome"]."</td>"."<td>".$gerente["dataNasc"]."</td>"."<td>".$gerente["sexo"]."</td>"."<td>".$gerente["email"]."</td><td>$editar</td>"."<td>$excluir</td>"."</tr>";
+                    }
                 }
             } catch(Exception $e){
                 print("Erro ...<br>".$e->getMessage());
